24Jan
COMBINATION SUM PROBLEMS

You are given an array of distinct intergers and you have to tell how many different ways to selecting the elements from the array are there such that the sum of chosen elements is equal to the target sum tar. Two ways are considered the same if contents of both ways are same for every index. example {1,2,3,1} and {1,2,3,1} are same while {1,2,3,4} and {4,3,2,1} are differents.

Detailed explanation ( Input/output format, Notes, Constraints, Images )
Sample Input 1 :
2
3 5
1 2 5
2 3
1 2
Sample Output 1 :
9
3
Explanation For Sample Output 1:
For the first test case, the number of possible ways will be
(1,1,1,1,1)
(1,1,1,2)
(1,2,1,1)
(2,1,1,1)
(1,1,2,1)
(2,2,1)
(1,2,2)
(2,1,2)
(5)

For the second test case, the number of ways will be
(1,1,1)
(1,2)
(2,1)
Here you can see we have considered (1,2) and (2,1) in 2 different ways.
Sample Input 2 :
2
3 4
12 1 3
2 41
2 34
Sample Output 2 :
3
0